v1_crypto/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pytz==2024.1
Flask==3.0.3
libsql-experimental==0.0.34
schedule==1.2.1
sendgrid==6.11.0
